Moved save_imgs function to utils.py
