+++ src/class_cvae/verify_latent_effectiveness.py
@@ -15,77 +15,12 @@ from PIL import Image, ImageDraw, ImageFont
 import matplotlib.pyplot as plt
 
 from models import Encoder, Decoder, Classifier, ImageClassifier
-from utils import create_img_from_text
+from utils import create_img_from_text, save_imgs
 
 def load_data():
     test_dset = MNIST(root="data", train=False, transform=ToTensor())
 
     return test_dset
-
-def save_imgs(reals, fakes, confs, org_confs, output):
-    reals = reals.cpu().detach().numpy()
-    fakes = fakes.cpu().detach().numpy()
-
-    reals = np.transpose(reals, (0, 2, 3, 1)) * 255
-    fakes = np.transpose(fakes, (0, 2, 3, 1)) * 255
-    diffs = (reals - fakes)
-    diffs_pos = np.copy(diffs)
-    diffs_pos[diffs_pos < 0] = 0
-    diffs_neg = np.copy(diffs)
-    diffs_neg[diffs_neg > 0] = 0
-    diffs_neg *= -1
-
-    final = None
-    for i in range(len(reals)):
-        if final is None:
-            final = create_img_from_text(reals.shape[1], 14, f"{int(round(org_confs[i].item(), 2)*100)}%")
-        else:
-            final = np.concatenate((final, create_img_from_text(reals.shape[1], 14, f"{int(round(org_confs[i].item(), 2)*100)}%")), axis=1)
-    
-    tmp = None
-    for i in range(len(reals)):
-        if tmp is None:
-            tmp = create_img_from_text(reals.shape[1], 14, f"{int(round(confs[i].item(), 2)*100)}%")
-        else:
-            tmp = np.concatenate((tmp, create_img_from_text(reals.shape[1], 14, f"{int(round(confs[i].item(), 2)*100)}%")), axis=1)
-    final = np.concatenate((final, tmp), axis=0)[:, :, :1].astype(np.uint8)
-    
-    tmp = None
-    for img in reals:
-        if tmp is None:
-            tmp = img
-        else:
-            tmp = np.concatenate((tmp, img), axis=1)
-    final = np.concatenate((final, tmp), axis=0)[:, :, :1].astype(np.uint8)
-
-    tmp = None
-    for img in fakes:
-        if tmp is None:
-            tmp = img
-        else:
-            tmp = np.concatenate((tmp, img), axis=1)
-
-    final = np.concatenate((final, tmp), axis=0)[:, :, :1].astype(np.uint8)
-    
-    tmp = None
-    for img in diffs_pos:
-        if tmp is None:
-            tmp = img
-        else:
-            tmp = np.concatenate((tmp, img), axis=1)
-
-    final = np.concatenate((final, tmp), axis=0)[:, :, :1].astype(np.uint8)
-    
-    tmp = None
-    for img in diffs_neg:
-        if tmp is None:
-            tmp = img
-        else:
-            tmp = np.concatenate((tmp, img), axis=1)
-
-    final = np.concatenate((final, tmp), axis=0)[:, :, :1].astype(np.uint8)
-
-    Image.fromarray(final[:, :, 0]).save(output)
 
 def get_args():
     parser = ArgumentParser()
